+++ src/Features/Core/Portable/NavigationBar/IRemoteNavigationBarItemService.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
 using System.Collections.Immutable;
 using System.Runtime.Serialization;
 using System.Threading;
@@ -46,7 +45,7 @@ namespace Microsoft.CodeAnalysis.NavigationBar
         public readonly SymbolKey? NavigationSymbolId;
 
         [DataMember(Order = 9)]
-        public readonly int? NavigationSymbolIndex;
+        public readonly int NavigationSymbolIndex;
 
         // Set for all GenerateCode kinds.
 
@@ -75,7 +74,7 @@ namespace Microsoft.CodeAnalysis.NavigationBar
             ImmutableArray<SerializableNavigationBarItem> childItems,
             ImmutableArray<TextSpan> spans,
             SymbolKey? navigationSymbolId,
-            int? navigationSymbolIndex,
+            int navigationSymbolIndex,
             SymbolKey? destinationTypeSymbolKey,
             string? containerName,
             SymbolKey? eventSymbolKey,
@@ -113,21 +112,21 @@ namespace Microsoft.CodeAnalysis.NavigationBar
             => values.SelectAsArray(v => v.Dehydrate());
 
         public static SerializableNavigationBarItem ActionlessItem(string text, Glyph glyph, ImmutableArray<TextSpan> spans, ImmutableArray<SerializableNavigationBarItem> childItems = default, int indent = 0, bool bolded = false, bool grayed = false)
-            => new(RoslynNavigationBarItemKind.Actionless, text, glyph, bolded, grayed, indent, childItems, spans, null, null, null, null, null, null);
+            => new(RoslynNavigationBarItemKind.Actionless, text, glyph, bolded, grayed, indent, childItems, spans, null, 0, null, null, null, null);
 
-        public static SerializableNavigationBarItem SymbolItem(string text, Glyph glyph, ImmutableArray<TextSpan> spans, SymbolKey navigationSymbolId, int? navigationSymbolIndex, ImmutableArray<SerializableNavigationBarItem> childItems = default, int indent = 0, bool bolded = false, bool grayed = false)
+        public static SerializableNavigationBarItem SymbolItem(string text, Glyph glyph, ImmutableArray<TextSpan> spans, SymbolKey navigationSymbolId, int navigationSymbolIndex, ImmutableArray<SerializableNavigationBarItem> childItems = default, int indent = 0, bool bolded = false, bool grayed = false)
             => new(RoslynNavigationBarItemKind.Symbol, text, glyph, bolded, grayed, indent, childItems, spans, navigationSymbolId, navigationSymbolIndex, null, null, null, null);
 
         public static SerializableNavigationBarItem GenerateFinalizer(string text, SymbolKey destinationTypeSymbolKey)
-            => new(RoslynNavigationBarItemKind.GenerateFinalizer, text, Glyph.MethodProtected, bolded: false, grayed: false, indent: 0, default, default, null, null, destinationTypeSymbolKey, null, null, null);
+            => new(RoslynNavigationBarItemKind.GenerateFinalizer, text, Glyph.MethodProtected, bolded: false, grayed: false, indent: 0, default, default, null, 0, destinationTypeSymbolKey, null, null, null);
 
         public static SerializableNavigationBarItem GenerateEventHandler(string eventName, Glyph glyph, string containerName, SymbolKey eventSymbolKey, SymbolKey destinationTypeSymbolKey)
-            => new(RoslynNavigationBarItemKind.GenerateEventHandler, eventName, glyph, bolded: false, grayed: false, indent: 0, default, default, null, null, destinationTypeSymbolKey, containerName, eventSymbolKey, null);
+            => new(RoslynNavigationBarItemKind.GenerateEventHandler, eventName, glyph, bolded: false, grayed: false, indent: 0, default, default, null, 0, destinationTypeSymbolKey, containerName, eventSymbolKey, null);
 
         public static SerializableNavigationBarItem GenerateMethod(string text, Glyph glyph, SymbolKey destinationTypeSymbolId, SymbolKey methodToReplicateSymbolId)
-            => new(RoslynNavigationBarItemKind.GenerateMethod, text, glyph, bolded: false, grayed: false, indent: 0, default, default, null, null, destinationTypeSymbolId, null, null, methodToReplicateSymbolId);
+            => new(RoslynNavigationBarItemKind.GenerateMethod, text, glyph, bolded: false, grayed: false, indent: 0, default, default, null, 0, destinationTypeSymbolId, null, null, methodToReplicateSymbolId);
 
         public static SerializableNavigationBarItem GenerateDefaultConstructor(string text, SymbolKey destinationTypeSymbolKey)
-            => new(RoslynNavigationBarItemKind.GenerateDefaultConstructor, text, Glyph.MethodPublic, bolded: false, grayed: false, indent: 0, default, default, null, null, destinationTypeSymbolKey, null, null, null);
+            => new(RoslynNavigationBarItemKind.GenerateDefaultConstructor, text, Glyph.MethodPublic, bolded: false, grayed: false, indent: 0, default, default, null, 0, destinationTypeSymbolKey, null, null, null);
     }
 }
